move test_additional_properties_is_true to TestSpecBackwardCompatibility
